requirements.txt
# LangChain Core
langchain
langchain-core

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai
langchain-community
langchain-text-splitters
langchain-chroma

# Environment Variable Management
python-dotenv
pydantic

# Machine Learning Utilities
numpy
scikit-learn
streamlit
sentence-transformers langchain-huggingface
sentence-transformers

# API Backend
fastapi
uvicorn

# Frontend HTTP Client
requests

beautifulsoup4